Dial behavior and end result bubbling
Dials now only spin while your finger is on it, also you do not have to
click on it at the start of the drag to do control it, just drag onto
it.

End results of pointer events now bubble up to t2c.js where it is easy
to respond to.